Move system prompt to a global option, and agents too
This means sub-clients reuse the same system-level prompt (a spike). And they can all call other agents too (probably not a great idea due to reentrancy).

Added lazy support for ChatOptions since the scheduler uses it and needs it lazy therefore, to avoid a cycle in DI (since Program.cs now adds all agents as tools).

In the history we now keep user messages as "User said..." since otherwise everytime we get responses for new messages, the older ones would be executed *again*. We need to shield from this somehow in a more robust way of course.